refactor(data): Extract ImageDetail, VideoDetail to core:model for Module Responsibility Alignment
