+++ core/data/src/main/java/com/wei/picquest/core/data/model/ImageDetail.kt
@@ -1,32 +1,7 @@
 package com.wei.picquest.core.data.model
 
+import com.wei.picquest.core.model.data.ImageDetail
 import com.wei.picquest.core.network.model.NetworkImageDetail
-
-data class ImageDetail(
-    val id: Int,
-    val pageURL: String,
-    val type: String,
-    val tags: String,
-    val previewURL: String,
-    val previewWidth: Int,
-    val previewHeight: Int,
-    val webformatURL: String,
-    val webformatWidth: Int,
-    val webformatHeight: Int,
-    val largeImageURL: String,
-    val imageWidth: Int,
-    val imageHeight: Int,
-    val imageSize: Long,
-    val views: Int,
-    val downloads: Int,
-    val likes: Int,
-    val comments: Int,
-    val userId: Int,
-    val user: String,
-    val userImageURL: String,
-) {
-    val aspectRatio get() = imageWidth.toFloat() / imageHeight.toFloat()
-}
 
 fun NetworkImageDetail.asExternalModel() = ImageDetail(
     id = this.id,
