+++ core/data/src/main/java/com/wei/picquest/core/data/repository/SearchImagesRepository.kt
@@ -1,7 +1,7 @@
 package com.wei.picquest.core.data.repository
 
 import androidx.paging.PagingData
-import com.wei.picquest.core.data.model.ImageDetail
+import com.wei.picquest.core.model.data.ImageDetail
 import kotlinx.coroutines.flow.Flow
 
 interface SearchImagesRepository {
