+++ core/model/src/main/java/com/wei/picquest/core/model/data/ImageDetail.kt
@@ -1,0 +1,27 @@
+package com.wei.picquest.core.model.data
+
+data class ImageDetail(
+    val id: Int,
+    val pageURL: String,
+    val type: String,
+    val tags: String,
+    val previewURL: String,
+    val previewWidth: Int,
+    val previewHeight: Int,
+    val webformatURL: String,
+    val webformatWidth: Int,
+    val webformatHeight: Int,
+    val largeImageURL: String,
+    val imageWidth: Int,
+    val imageHeight: Int,
+    val imageSize: Long,
+    val views: Int,
+    val downloads: Int,
+    val likes: Int,
+    val comments: Int,
+    val userId: Int,
+    val user: String,
+    val userImageURL: String,
+) {
+    val aspectRatio get() = imageWidth.toFloat() / imageHeight.toFloat()
+}
