+++ feature/photo/src/main/java/com/wei/picquest/feature/photo/photolibrary/PhotoLibraryViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.wei.picquest.core.base.BaseViewModel
-import com.wei.picquest.core.data.model.ImageDetail
 import com.wei.picquest.core.data.repository.SearchImagesRepository
+import com.wei.picquest.core.model.data.ImageDetail
 import com.wei.picquest.feature.photo.photolibrary.navigation.PhotoLibraryArgs
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
