+++ feature/video/src/main/java/com/wei/picquest/feature/video/videolibrary/VideoLibraryScreen.kt
@@ -63,15 +63,15 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
-import com.wei.picquest.core.data.model.VideoDetail
-import com.wei.picquest.core.data.model.VideoDetailSize
-import com.wei.picquest.core.data.model.VideoStreams
 import com.wei.picquest.core.designsystem.component.ThemePreviews
 import com.wei.picquest.core.designsystem.component.coilImagePainter
 import com.wei.picquest.core.designsystem.icon.PqIcons
 import com.wei.picquest.core.designsystem.theme.PqTheme
 import com.wei.picquest.core.designsystem.theme.SPACING_MEDIUM
 import com.wei.picquest.core.designsystem.theme.SPACING_SMALL
+import com.wei.picquest.core.model.data.VideoDetail
+import com.wei.picquest.core.model.data.VideoDetailSize
+import com.wei.picquest.core.model.data.VideoStreams
 import com.wei.picquest.core.pip.enterPictureInPicture
 import com.wei.picquest.core.pip.isInPictureInPictureMode
 import com.wei.picquest.core.pip.updatedPipParams
